# docs/openapi3.yaml
openapi: 3.0.0
info:
  version: 1.7.1
  title: Predator
  x-logo:
    url: logo.png
    backgroundColor: white
  description: >
    # Overview


    Predator is an open-source performance framework. It manages the entire

    lifecycle of load testing a server, from creating a test file, running

    scheduled and on-demand tests, and finally viewing the test results.

    Developers who want to load their system end-to-end or test a specific

    service will benefit from this framework. Bootstrapped with a user-friendly

    UI alongside a simple REST API, Predator simplifies the performance testing

    regime.


    For an introduction to Predator and its concepts, see the <a
    href="https://www.predator.dev" target="_blank">Predator Website</a>.
...
paths:
...
  /v1/tests:
    parameters:
      - $ref: '#/components/parameters/context_id'
    post:
      operationId: create-a-test
      tags:
        - Tests
      summary: Create a Test
      description: Create a new test.
      parameters:
        - in: query
          name: filter
          description: is filter and by what property
          required: false
          schema:
            type: string
            enum:
              - is_favorite
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/create_test_response'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/test'
        description: The test to add
        required: true
    get:
      operationId: retrieve-all-tests
      tags:
        - Tests
      summary: Retrieve all Tests
      description: Retrieve all available tests.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/tests_response'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
...
  /v1/jobs:
    parameters:
      - $ref: '#/components/parameters/context_id'
    post:
      operationId: create-a-job
      tags:
        - Jobs
      summary: Create a Job
      description: Create a new job that will trigger a test run.
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/create_job_response'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/job'
        description: The job to add
        required: true
    get:
      operationId: retrieve-all-jobs
      tags:
        - Jobs
      summary: Retrieve all Jobs
      description: Retrieve the details of all available jobs.
      parameters:
        - in: query
          name: one_time
          description: Should return the one time tests (no cron expression)
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/job_response'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
...
  /v1/config:
    parameters:
      - $ref: '#/components/parameters/context_id'
    get:
      operationId: retrieve-configurations
      tags:
        - Configuration
      summary: Retrieve Configuration Settings
      description: Retrieve all Predator configuration settings.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/config_body'
        '500':
          description: Internal server error
    put:
      operationId: update-configurations
      tags:
        - Configuration
      summary: Update Configuration Settings
      description: >-
        Update existing configuration settings. This will override environment
        variables.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/config_body'
        description: The config data to add
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/config_body'
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
...
  /v1/processors:
    parameters:
      - $ref: '#/components/parameters/context_id'
    post:
      operationId: create-processor-file
      tags:
        - Processors
      summary: Create processor file
      description: >-
        Create a new processor file which will have javascript code that can be
        executed by tests.
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/processor'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/processor'
        description: The processor file to create
        required: true
    get:
      operationId: retrieve-all-processor-files
      tags:
        - Processors
      summary: Retrieve all processors files
      description: Retrieve the details of all processor files.
      parameters:
        - in: query
          name: from
          description: From which result to start retrieve data
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          description: Max results to return from this query ( Limit to 100 )
          required: false
          schema:
            type: integer
            maximum: 200
            default: 100
        - in: query
          name: exclude
          description: fields to exclude from response
          required: false
          schema:
            type: array
            items:
              type: string
              enum:
                - javascript
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/processor'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
...
  /v1/webhooks:
    parameters:
      - $ref: '#/components/parameters/context_id'
    get:
      operationId: retrieve-webhooks
      tags:
        - Webhooks
      summary: Retrieve webhooks
      description: Retrieve all webhooks.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/webhook'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
    post:
      operationId: create-a-webhook
      tags:
        - Webhooks
      summary: Create a Webhook
      description: Create a new Webhook.
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/webhook'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/webhook'
        description: The webhook to add
        required: true
...
  /v1/files:
    parameters:
      - $ref: '#/components/parameters/context_id'
    post:
      operationId: upload-file
      tags:
        - Files
      summary: Create file
      description: Upload a file which can be used by Predator tests like csv.
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/upload_file_response'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '413':
          description: Payload Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                csv:
                  type: string
                  format: binary
        description: The file to create upload
        required: true
...
  /v1/chaos-experiments:
    parameters:
      - $ref: '#/components/parameters/context_id'
    post:
      operationId: create-chaos-experiment
      tags:
        - Chaos Experiments
      summary: Create chaos file
      description: Create a chaos experiment.
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chaosExperiment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chaosExperiment'
        description: The chaos experiment to create
        required: true
    get:
      operationId: retrieve-all-choos-experiments
      tags:
        - Chaos Experiments
      summary: Retrieve all chaos experiments
      description: Retrieve all chaos experiments
      parameters:
        - in: query
          name: from
          description: From which result to start retrieve data - pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          description: Max results to return from this query ( Limit to 100 )
          required: false
          schema:
            type: integer
            maximum: 200
            default: 100
        - in: query
          name: template
          description: filter to get templates only
          required: false
          schema:
            type: boolean
            example: true
        - in: query
          name: exclude
          description: fields to exclude from response
          required: false
          schema:
            type: array
            items:
              type: string
              enum:
                - kubeObject
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/chaosExperiment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error_response'
...
components:
  parameters:
    context_id:
      in: header
      name: x-context-id
      description: The id of the user/organization context
      required: false
      schema:
        type: string
        example: team-mick
...
  schemas:
...
    test:
      type: object
      discriminator:
        propertyName: type
      oneOf:
        - $ref: '#/components/schemas/basic'
        - $ref: '#/components/schemas/dsl'
      required:
        - name
        - description
        - type
      properties:
        name:
          type: string
          minLength: 1
          description: The name of the test.
          example: Order from Pet Store
        processor_file_url:
          type: string
          format: uri
          example: http://www.example.com/file.txt
          description: Url to text file location , will be used by the processor
        csv_file_id:
          type: string
          format: uuid
          example: 8e83bca3-f04f-4d16-a324-b0d2959db7d4
          description: The id of the uploaded csv file resource
        description:
          type: string
          description: A description of the test.
          example: Performance test for placing orders from the pet store.
        processor_id:
          description: Unique processor identifier
          type: string
          format: uuid
        is_favorite:
          type: boolean
          description: Test has been marked favorite
        type:
          $ref: '#/components/schemas/test_type'
    test_type:
      type: string
      description: >
        The type of test. Can be one of the following:

        * `basic`: Tests of this type cannot reuse request templates (DSL
        definition). All request specifications must be directly included in the
        test itself.
          This test type is identical to the type of tests you can create through the Predator UI.
        * `dsl`: Tests of this type reuse request templates based on previously
        created DSL definitons.

        enum:

        - dsl

        - basic

        example: dsl
...
    dsl:
      description: A test that is made of scenarios base on domain specific language
      required:
        - scenarios
      properties:
        scenarios:
          description: >-
            A scenario is a sequence of HTTP requests aimed to test the
            performance of specific functionality.
          type: array
          items:
            $ref: '#/components/schemas/scenario'
...
    basic:
      description: A test that is made of artillery json
      required:
        - artillery_test
      properties:
        artillery_test:
          type: object
    test_id:
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          description: The id of the test.
    create_test_response:
      allOf:
        - $ref: '#/components/schemas/test_id'
      required:
        - revision_id
      properties:
        revision_id:
          type: string
          format: uuid
          description: >
            The test's revision ID.


            When a test is created, it is assigned a revision ID. If you update
            a test, then it's revision ID changes as well. This allows Predator
            to always relate an existing report to the test it was originally
            based on.
    tests_response:
      type: object
      discriminator:
        propertyName: type
        mapping:
          dsl: '#/components/schemas/dsl_test_response'
          basic: '#/components/schemas/basic_test_response'
      required:
        - type
        - name
        - description
        - updated_at
      properties:
        type:
          $ref: '#/components/schemas/test_type'
        updated_at:
          type: string
          format: date-time
          description: The date and time that the test was created.
        name:
          type: string
          description: The name of the test.
        description:
          type: string
          description: The description of the test.
        file_id:
          type: string
          description: File id.
        processor_id:
          description: Unique processor identifier
          type: string
          format: uuid
...
    error_response:
      required:
        - message
      properties:
        message:
          type: string
...
    job:
      discriminator:
        propertyName: type
      oneOf:
        - $ref: '#/components/schemas/load_test'
        - $ref: '#/components/schemas/functional_test'
      required:
        - type
        - test_id
        - duration
      properties:
        run_immediately:
          type: boolean
          description: >-
            Determines if the test will be executed immediately when the job is
            created.
          example: true
        type:
          type: string
          enum:
            - load_test
            - functional_test
          description: The type of the job to run.
          example: load_test
        experiments:
          $ref: '#/components/schemas/job_experiments'
...
    job_response:
      allOf:
        - $ref: '#/components/schemas/job'
        - properties:
            id:
              type: string
              pattern: >-
                ^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$
                description: The job id.
    create_job_response:
      allOf:
        - $ref: '#/components/schemas/job_response'
        - properties:
            report_id:
              type: string
              description: The run id. Only shown if the job is ran immediately.
    processor:
      required:
        - name
        - description
        - javascript
      properties:
        id:
          description: Unique processor identifier
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          description: The name of the processor.
          example: Custom javascript for logging
        description:
          type: string
          description: A description of the processor.
          example: logs every error (5xx).
        exported_functions:
          type: array
          readOnly: true
          description: Names of all exported function in the javascript file
          items:
            type: string
          example: generateNameBeforeScenario
        updated_at:
          type: string
          format: date-time
          description: The date and time that the procesor file was updated.
          readOnly: true
        created_at:
          type: string
          format: date-time
          description: The date and time that the procesor file was created.
          readOnly: true
        javascript:
          type: string
          description: Raw javascript that will be used by the processor resource.
          example: >-
            module.exports.generateNameBeforeScenario = (context, ee, next) => {
            context.name = Date.now() + '_mick' }
...
    config_body:
      additionalProperties: false
      properties:
        grafana_url:
          type: string
          description: The url for grafana.
        internal_address:
          type: string
          description: The local ip address of your machine.
        runner_docker_image:
          type: string
          description: The predator-runner docker image that will run the test.
        job_platform:
          type: string
          description: >-
            Type of platform using to run predator
            (METRONOME,KUBERNETES,DOCKER,AWS_FARGATE)
        runner_cpu:
          minimum: 0
          type: number
          description: Number of CPU use by the each runner.
        runner_memory:
          type: number
          minimum: 128
          description: Max memory to use by each runner.
        allow_insecure_tls:
          type: boolean
          description: >-
            If true, don't fail requests on unverified server certificate
            errors.
        interval_cleanup_finished_containers_ms:
          type: number
          minimum: 0
          description: >-
            Interval (in ms) to search and delete finished tests containers.
            Value of 0 means no auto clearing enabled.
        chaos_mesh_enabled:
          type: boolean
          description: If true, allow chaos mesh experiments
        kafka_brokers:
          type: string
          description: 'Comma-separated Kafka broker list for kafka load tests (host:port,host:port)'
        kafka_ssl:
          type: boolean
          description: Use SSL when connecting to Kafka
        kafka_sasl_mechanism:
          type: string
          enum:
            - plain
            - scram-sha-256
            - scram-sha-512
          description: SASL mechanism for Kafka authentication
        kafka_sasl_username:
          type: string
          description: SASL username for Kafka
        kafka_sasl_password:
          type: string
          description: SASL password for Kafka
        metrics_plugin_name:
          type: string
          enum:
            - prometheus
          description: Metrics to use, in case one or both metrics are configured.
        default_email_address:
          type: string
          description: >-
            Default email to send final report to, address can be configured
            only by using API.
        minimum_wait_for_delayed_report_status_update_in_ms:
          minimum: 0
          type: number
          description: >-
            The minimum of time waiting for runner to report before the test
            considered as finished. configured only by using API.
        delay_runner_ms:
          type: number
          minimum: 0
          description: >-
            Delay before the runners will start the test, useful for enviroments
            like kubernetes with sidecars that should be initialized before the
            runner starts
        prometheus_metrics:
          additionalProperties: false
          type: object
          description: prometheus configuration.
          required:
            - push_gateway_url
          properties:
            push_gateway_url:
              type: string
              description: Url of push gateway.
            buckets_sizes:
              type: array
              items:
                type: number
              description: Bucket sizes use to configure prometheus.
            labels:
              description: Labels which will be passed to prometheus push gateway.
              type: object
              additionalProperties:
                type: string
        smtp_server:
          additionalProperties: false
          type: object
          description: Simple mail transfer protocol.
          required:
            - from
            - host
            - port
            - username
            - password
            - timeout
          properties:
            from:
              type: string
              format: email
              description: the 'from' email address that will be used to send emails
            host:
              type: string
            port:
              type: number
            username:
              type: string
            password:
              type: string
            timeout:
              description: timout to smtp server in milliseconds
              type: number
        benchmark_threshold:
          type: number
          description: >-
            Minimum acceptable score of tests, if a score is less than this
            value, a webhook will be sent to the threshold webhook url.
          minimum: 0
          maximum: 100
        benchmark_weights:
          additionalProperties: false
          type: object
          description: benchmark_weights
          required:
            - percentile_ninety_five
            - percentile_fifty
            - server_errors_ratio
            - client_errors_ratio
            - rps
          properties:
            percentile_ninety_five:
              description: Percentage of the score affected by p95 results.
              allOf:
                - $ref: '#/components/schemas/benchmark_weights'
            percentile_fifty:
              description: Percentage of the score affected by median results.
              allOf:
                - $ref: '#/components/schemas/benchmark_weights'
            server_errors_ratio:
              description: Percentage of the score affected by server errors ratio.
              allOf:
                - $ref: '#/components/schemas/benchmark_weights'
            client_errors_ratio:
              description: Percentage of the score affected by client errors ratio.
              allOf:
                - $ref: '#/components/schemas/benchmark_weights'
            rps:
              description: Percentage of the score affected by requests per second results.
              allOf:
                - $ref: '#/components/schemas/benchmark_weights'
        custom_runner_definition:
          description: >-
            custom json that will be merged with the kubernetes/metronome/aws
            fargate predator runner job definition
          type: object
        streaming_excluded_attributes:
          description: >-
            string of attributes to remove from streamed messages with ',' as
            delimeter
          type: string
...
    upload_file_response:
      type: object
      required:
        - id
        - filename
      properties:
        id:
          type: string
          pattern: uuid
          description: the id of the file
        filename:
          type: string
          description: the name of the file
...
    functional_test:
      allOf:
        - $ref: '#/components/schemas/job_update'
        - type: object
      type: object
      required:
        - arrival_count
      properties:
        arrival_count:
          type: number
          minimum: 1
          description: The number of times per second that the test scenarios will run.
    webhook:
      type: object
      required:
        - name
        - url
        - events
        - format_type
      properties:
        id:
          description: Unique webhook identifier
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          description: Webhook name
        url:
          type: string
          description: Webhook url to post events
        events:
          description: list of events which will trigger the webhook
          type: array
          uniqueItems: true
          minItems: 1
          items:
            $ref: '#/components/schemas/webhooks_types'
        format_type:
          $ref: '#/components/schemas/webhook_format_types'
        global:
          type: boolean
          description: >-
            indicates whether the webhook should be applied globally(over all
            jobs)
...
    webhook_format_types:
      type: string
      enum:
        - slack
        - json
        - teams
        - discord
...
    chaosExperiment:
      required:
        - name
        - kubeObject
      properties:
        id:
          description: Unique experiment identifier
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          description: The name of the experiment.
          example: Custom javascript for logging
        kubeObject:
          $ref: '#/components/schemas/kubeObject'
        updated_at:
          type: string
          format: date-time
          description: The date and time that the chaos experiment file was updated.
          readOnly: true
        created_at:
          type: string
          format: date-time
          description: The date and time that the chaos experiment was created.
          readOnly: true
...